+++ src/AgreementSearch.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, Icon } from '@folio/stripes/components';
+import contains from 'dom-helpers/query/contains';
 import Modal from './Modal';
 
 const triggerId = 'find-agreement-trigger';
@@ -10,26 +11,39 @@ export default class AgreementSearch extends React.Component {
     renderTrigger: PropTypes.func,
   };
 
-  state = {
-    open: false,
-  };
+  constructor(props) {
+    super(props);
+
+    this.modalRef = React.createRef();
+    this.modalTrigger = React.createRef();
+    this.state = {
+      open: false,
+    };
+  }
 
   openModal = () => {
     this.setState({ open: true });
   }
 
   closeModal = () => {
-    this.setState({ open: false });
+    this.setState({ open: false }, () => {
+      if (this.modalRef.current && this.modalTrigger.current) {
+        if (contains(this.modalRef.current, document.activeElement)) {
+          this.modalTrigger.current.focus();
+        }
+      }
+    });
   }
 
   renderDefaultTrigger() {
     return (
       <Button
-        id={triggerId}
+        buttonRef={this.modalTrigger}
         buttonStyle="primary noRightRadius"
+        id={triggerId}
         onClick={this.openModal}
       >
-        <Icon icon="search" color="#fff" />
+        <Icon color="#fff" icon="search" />
       </Button>
     );
   }
@@ -41,6 +55,7 @@ export default class AgreementSearch extends React.Component {
 
     return renderTrigger
       ? renderTrigger({
+        buttonRef: this.modalTrigger,
         id: triggerId,
         onClick: this.openModal,
       })
@@ -49,14 +64,15 @@ export default class AgreementSearch extends React.Component {
 
   render() {
     return (
-      <React.Fragment>
+      <>
         {this.renderTriggerButton()}
         <Modal
-          open={this.state.open}
+          modalRef={this.modalRef}
           onClose={this.closeModal}
+          open={this.state.open}
           {...this.props}
         />
-      </React.Fragment>
+      </>
 
     );
   }