+++ src/Container.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { get } from 'lodash';
 
 import { StripesConnectedSource } from '@folio/stripes/smart-components';
-import { getSASParams } from '@folio/stripes-erm-components';
+import { generateQueryParams, preventResourceRefresh } from '@folio/stripes-erm-components';
 
 import View from './View';
 
 const INITIAL_RESULT_COUNT = 100;
 const RESULT_COUNT_INCREMENT = 100;
+const RECORDS_PER_REQUEST = 100;
 
 export default class Container extends React.Component {
   static manifest = Object.freeze({
@@ -19,35 +19,54 @@ export default class Container extends React.Component {
       recordsRequired: '%{resultCount}',
       perRequest: 100,
       limitParam: 'perPage',
-      params: getSASParams({
-        searchKey: 'name',
+      params: generateQueryParams({
+        searchKey: 'name,alternateNames.name,description',
         filterKeys: {
+          agreementStatus: 'agreementStatus.value',
+          contactRole: 'contacts.role',
+          contacts: 'contacts.user',
+          isPerpetual: 'isPerpetual.value',
+          renewalPriority: 'renewalPriority.value',
           orgs: 'orgs.org',
-          role: 'orgs.role',
-          tags: 'tags.value',
+          role: 'orgs.roles.role',
+          tags: 'tags.label',
         },
-        queryGetter: r => r.agreementSearchParams,
+        sortKeys: {
+          agreementStatus: 'agreementStatus.label',
+        },
       }),
+    },
+    contactRoleValues: {
+      type: 'okapi',
+      path: 'erm/refdata/InternalContact/role',
+      limitParam: 'perPage',
+      perRequest: RECORDS_PER_REQUEST,
+      shouldRefresh: () => false,
     },
     agreementStatusValues: {
       type: 'okapi',
-      path: 'erm/refdataValues/SubscriptionAgreement/agreementStatus',
+      path: 'erm/refdata/SubscriptionAgreement/agreementStatus',
       shouldRefresh: () => false,
     },
     renewalPriorityValues: {
       type: 'okapi',
-      path: 'erm/refdataValues/SubscriptionAgreement/renewalPriority',
+      path: 'erm/refdata/SubscriptionAgreement/renewalPriority',
       shouldRefresh: () => false,
     },
     isPerpetualValues: {
       type: 'okapi',
-      path: 'erm/refdataValues/SubscriptionAgreement/isPerpetual',
+      path: 'erm/refdata/SubscriptionAgreement/isPerpetual',
       shouldRefresh: () => false,
     },
     orgRoleValues: {
       type: 'okapi',
-      path: 'erm/refdataValues/SubscriptionAgreementOrg/role',
+      path: 'erm/refdata/SubscriptionAgreementOrg/role',
       shouldRefresh: () => false,
+    },
+    supplementaryProperties: {
+      type: 'okapi',
+      path: 'erm/custprops',
+      shouldRefresh: preventResourceRefresh({ 'agreement': ['DELETE'] }),
     },
     tagsValues: {
       type: 'okapi',
@@ -60,6 +79,7 @@ export default class Container extends React.Component {
         sort: 'name',
       }
     },
+    query: { initialValue: {} },
     resultCount: { initialValue: INITIAL_RESULT_COUNT },
   });
 
@@ -86,8 +106,8 @@ export default class Container extends React.Component {
       this.searchField.current.focus();
     }
 
-    this.props.mutator.agreementSearchParams.update({
-      filters: 'agreementStatus.Active,agreementStatus.Draft,agreementStatus.In negotiation,agreementStatus.Requested',
+    this.props.mutator.query.update({
+      filters: 'agreementStatus.active,agreementStatus.draft,agreementStatus.in_negotiation,agreementStatus.requested',
     });
   }
 
@@ -97,16 +117,12 @@ export default class Container extends React.Component {
     }
   }
 
-  querySetter = ({ nsValues, state }) => {
-    if (/reset/.test(state.changeType)) {
-      this.props.mutator.agreementSearchParams.replace(nsValues);
-    } else {
-      this.props.mutator.agreementSearchParams.update(nsValues);
-    }
+  querySetter = ({ nsValues }) => {
+    this.props.mutator.query.update(nsValues);
   }
 
   queryGetter = () => {
-    return get(this.props.resources, 'agreementSearchParams', {});
+    return this.props?.resources?.query ?? {};
   }
 
   render() {
@@ -119,12 +135,14 @@ export default class Container extends React.Component {
     return (
       <View
         data={{
-          agreements: get(resources, 'agreements.records', []),
-          agreementStatusValues: get(resources, 'agreementStatusValues.records', []),
-          renewalPriorityValues: get(resources, 'renewalPriorityValues.records', []),
-          isPerpetualValues: get(resources, 'isPerpetualValues.records', []),
-          orgRoleValues: get(resources, 'orgRoleValues.records', []),
-          tagsValues: get(resources, 'tagsValues.records', []),
+          agreements: resources?.agreements?.records ?? [],
+          agreementStatusValues: resources?.agreementStatusValues?.records ?? [],
+          renewalPriorityValues: resources?.renewalPriorityValues?.records ?? [],
+          isPerpetualValues: resources?.isPerpetualValues?.records ?? [],
+          orgRoleValues: resources?.orgRoleValues?.records ?? [],
+          contactRoleValues: resources?.contactRoleValues?.records ?? [],
+          tagsValues: resources?.tagsValues?.records ?? [],
+          supplementaryProperties: resources?.supplementaryProperties?.records ?? []
         }}
         onNeedMoreData={this.handleNeedMoreData}
         onSelectRow={onSelectRow}