+++ src/Filters.js
@@ -1,10 +1,17 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 
 import { Accordion, AccordionSet, FilterAccordionHeader, Selection } from '@folio/stripes/components';
+import { IfPermission } from '@folio/stripes/core';
 import { CheckboxFilter, MultiSelectionFilter } from '@folio/stripes/smart-components';
-import { OrganizationSelection } from '@folio/stripes-erm-components';
+import { CustomPropertyFilters, DateFilter, InternalContactSelection, OrganizationSelection } from '@folio/stripes-erm-components';
+
+const propTypes = {
+  activeFilters: PropTypes.object,
+  data: PropTypes.object.isRequired,
+  filterHandlers: PropTypes.object,
+};
 
 const FILTERS = [
   'agreementStatus',
@@ -13,58 +20,51 @@ const FILTERS = [
   'tags'
 ];
 
-export default class Filters extends React.Component {
-  static propTypes = {
-    activeFilters: PropTypes.object,
-    data: PropTypes.object.isRequired,
-    filterHandlers: PropTypes.object,
-  };
+export default function Filters({ activeFilters, data, filterHandlers }) {
+  const intl = useIntl();
 
-  static defaultProps = {
-    activeFilters: {}
-  };
-
-  state = {
+  const [filterState, setFilterState] = useState({
     agreementStatus: [],
     renewalPriority: [],
     isPerpetual: [],
-    tags: [],
-  }
+    tags: []
+  });
 
-  static getDerivedStateFromProps(props, state) {
+  useEffect(() => {
     const newState = {};
-
     FILTERS.forEach(filter => {
-      const values = props.data[`${filter}Values`] || [];
-      if (values.length !== state[filter].length) {
-        newState[filter] = values.map(({ label }) => ({ label, value: label }));
+      const values = data[`${filter}Values`];
+      if (values.length !== filterState[filter]?.length) {
+        newState[filter] = values;
       }
     });
 
-    if (Object.keys(newState).length) return newState;
+    if ((data?.tagsValues?.length ?? 0) !== filterState.tags?.length) {
+      newState.tags = data.tagsValues.map(({ label }) => ({ value: label, label }));
+    }
 
-    return null;
-  }
+    if (Object.keys(newState).length) {
+      setFilterState(prevState => ({ ...prevState, ...newState }));
+    }
+  }, [data, filterState]);
 
-  renderCheckboxFilter = (name, props) => {
-    const { activeFilters } = this.props;
+  const renderCheckboxFilter = (name, props) => {
     const groupFilters = activeFilters[name] || [];
-
     return (
       <Accordion
         displayClearButton={groupFilters.length > 0}
         header={FilterAccordionHeader}
         id={`filter-accordion-${name}`}
         label={<FormattedMessage id={`ui-plugin-find-agreement.prop.${name}`} />}
-        onClearFilter={() => { this.props.filterHandlers.clearGroup(name); }}
+        onClearFilter={() => { filterHandlers.clearGroup(name); }}
         separator={false}
         {...props}
       >
         <CheckboxFilter
-          dataOptions={this.state[name]}
+          dataOptions={filterState[name] || []}
           name={name}
           onChange={(group) => {
-            this.props.filterHandlers.state({
+            filterHandlers.state({
               ...activeFilters,
               [group.name]: group.values
             });
@@ -73,10 +73,9 @@ export default class Filters extends React.Component {
         />
       </Accordion>
     );
-  }
+  };
 
-  renderOrganizationFilter = () => {
-    const { activeFilters } = this.props;
+  const renderOrganizationFilter = () => {
     const orgFilters = activeFilters.orgs || [];
 
     return (
@@ -84,9 +83,10 @@ export default class Filters extends React.Component {
         closedByDefault
         displayClearButton={orgFilters.length > 0}
         header={FilterAccordionHeader}
+        id="organizations-filter"
         label={<FormattedMessage id="ui-plugin-find-agreement.prop.organizations" />}
         onClearFilter={() => {
-          this.props.filterHandlers.state({
+          filterHandlers.state({
             ...activeFilters,
             role: [],
             orgs: [],
@@ -97,24 +97,22 @@ export default class Filters extends React.Component {
         <OrganizationSelection
           input={{
             name: 'agreement-orgs-filter',
-            onChange: value => this.props.filterHandlers.state({ ...activeFilters, orgs: [value] }),
+            onChange: value => filterHandlers.state({ ...activeFilters, orgs: [value] }),
             value: orgFilters[0] || '',
           }}
         />
       </Accordion>
     );
-  }
+  };
 
-  renderOrganizationRoleFilter = () => {
-    const roles = this.props.data.orgRoleValues;
+  const renderOrganizationRoleFilter = () => {
+    const roles = data.orgRoleValues;
     // TODO: TEST USING THE VALUES GENERATED IN GDSFP
     const dataOptions = roles.map(role => ({
       value: role.id,
       label: role.label,
     }));
 
-    const { activeFilters } = this.props;
-    const orgFilters = activeFilters.orgs || [];
     const roleFilters = activeFilters.role || [];
 
     return (
@@ -122,58 +120,152 @@ export default class Filters extends React.Component {
         closedByDefault
         displayClearButton={roleFilters.length > 0}
         header={FilterAccordionHeader}
+        id="organization-role-filter"
         label={<FormattedMessage id="ui-plugin-find-agreement.prop.orgRole" />}
-        onClearFilter={() => { this.props.filterHandlers.clearGroup('role'); }}
+        onClearFilter={() => { filterHandlers.clearGroup('role'); }}
+        separator={false}
+      >
+        <FormattedMessage id="ui-plugin-find-agreement.organizations.selectRole">
+          {placeholder => (
+            <Selection
+              dataOptions={dataOptions}
+              onChange={value => filterHandlers.state({ ...activeFilters, role: [value] })}
+              placeholder={typeof placeholder === 'string' ? placeholder : placeholder[0]}
+              value={roleFilters[0] || ''}
+            />
+          )}
+        </FormattedMessage>
+      </Accordion>
+    );
+  };
+
+  const renderInternalContactFilter = () => {
+    const contactFilters = activeFilters.contacts || [];
+
+    return (
+      <IfPermission perm="users.collection.get">
+        <Accordion
+          closedByDefault
+          displayClearButton={contactFilters.length > 0}
+          header={FilterAccordionHeader}
+          id="internal-contacts-filter"
+          label={<FormattedMessage id="ui-plugin-find-agreement.agreements.internalContacts" />}
+          onClearFilter={() => filterHandlers.clearGroup('contacts')}
+          separator={false}
+        >
+          <InternalContactSelection
+            id="agreement-internal-contacts-filter"
+            input={{
+              name: 'agreement-contacts-filter',
+              onChange: value => filterHandlers.state({ ...activeFilters, contacts: [value] }),
+              value: contactFilters[0] || '',
+            }}
+            path="erm/contacts"
+          />
+        </Accordion>
+      </IfPermission>
+    );
+  };
+
+  const renderInternalContactRoleFilter = () => {
+    const contactRoles = data.contactRoleValues || [];
+    const dataOptions = contactRoles.map(contactRole => ({
+      value: contactRole.id,
+      label: contactRole.label,
+    }));
+    const contactRoleFilters = activeFilters.contactRole || [];
+    return (
+      <Accordion
+        closedByDefault
+        displayClearButton={contactRoleFilters.length > 0}
+        header={FilterAccordionHeader}
+        id="internal-contacts-role-filter"
+        label={<FormattedMessage id="ui-plugin-find-agreement.agreements.internalContactsRole" />}
+        onClearFilter={() => { filterHandlers.clearGroup('contactRole'); }}
         separator={false}
       >
         <Selection
           dataOptions={dataOptions}
-          disabled={orgFilters.length === 0}
-          value={roleFilters[0] || ''}
-          onChange={value => this.props.filterHandlers.state({ ...activeFilters, role: [value] })}
+          id="agreement-internal-contacts-role-filter"
+          onChange={value => filterHandlers.state({ ...activeFilters, contactRole: [value] })}
+          value={contactRoleFilters[0] || ''}
         />
       </Accordion>
     );
-  }
+  };
 
-  renderTagsFilter = () => {
+  const renderCustomPropertyFilters = () => {
+    return <CustomPropertyFilters
+      activeFilters={activeFilters}
+      customProperties={data.supplementaryProperties || []}
+      custPropName="supplementaryProperty"
+      filterHandlers={filterHandlers}
+    />;
+  };
+
+  const renderStartDateFilter = () => {
+    return <DateFilter
+      activeFilters={activeFilters}
+      filterHandlers={filterHandlers}
+      hideNoDateSetCheckbox
+      name="startDate"
+      resourceName={intl.formatMessage({ id: 'ui-plugin-find-agreement.agreements' }).toLowerCase()}
+    />;
+  };
+
+  const renderEndDateFilter = () => {
+    return <DateFilter
+      activeFilters={activeFilters}
+      filterHandlers={filterHandlers}
+      name="endDate"
+      resourceName={intl.formatMessage({ id: 'ui-plugin-find-agreement.agreements' }).toLowerCase()}
+    />;
+  };
+  const renderTagsFilter = () => {
     // const tags = get(this.props.data, 'tagValues.records', []);
     // TODO: TEST USING THE VALUES GENERATED IN GDSFP
     // const dataOptions = tags.map(({ label }) => ({ value: label, label }));
-    const { activeFilters } = this.props;
     const tagFilters = activeFilters.tags || [];
 
     return (
       <Accordion
         closedByDefault
-        id="clickable-tags-filter"
         displayClearButton={activeFilters.length > 0}
         header={FilterAccordionHeader}
+        id="clickable-tags-filter"
         label={<FormattedMessage id="ui-plugin-find-agreement.prop.tags" />}
-        onClearFilter={() => { this.props.filterHandlers.clearGroup('tags'); }}
+        onClearFilter={() => { filterHandlers.clearGroup('tags'); }}
         separator={false}
       >
         <MultiSelectionFilter
-          dataOptions={this.state.tags}
+          dataOptions={filterState.tags || []}
           id="tags-filter"
           name="tags"
-          onChange={e => this.props.filterHandlers.state({ ...activeFilters, tags: e.values })}
+          onChange={e => filterHandlers.state({ ...activeFilters, tags: e.values })}
           selectedValues={tagFilters}
         />
       </Accordion>
     );
-  }
+  };
 
-  render() {
-    return (
-      <AccordionSet>
-        {this.renderCheckboxFilter('agreementStatus')}
-        {this.renderCheckboxFilter('renewalPriority', { closedByDefault: true })}
-        {this.renderCheckboxFilter('isPerpetual', { closedByDefault: true })}
-        {this.renderOrganizationFilter()}
-        {this.renderOrganizationRoleFilter()}
-        {this.renderTagsFilter()}
-      </AccordionSet>
-    );
-  }
+  return (
+    <AccordionSet>
+      {renderCheckboxFilter('agreementStatus')}
+      {renderCheckboxFilter('renewalPriority', { closedByDefault: true })}
+      {renderCheckboxFilter('isPerpetual', { closedByDefault: true })}
+      {renderStartDateFilter()}
+      {renderEndDateFilter()}
+      {renderOrganizationFilter()}
+      {renderOrganizationRoleFilter()}
+      {renderInternalContactFilter()}
+      {renderInternalContactRoleFilter()}
+      {renderTagsFilter()}
+      {renderCustomPropertyFilters()}
+    </AccordionSet>
+  );
 }
+
+Filters.propTypes = propTypes;
+Filters.defaultProps = {
+  activeFilters: {}
+};