+++ CHANGELOG.md
@@ -1,4 +1,49 @@
 # Change history for ui-plugin-find-agreement
+## 8.2.0 In progress
+* ERM-2101 Replace babel-eslint with @babel/eslint-parser
+
+## 8.1.0 2022-03-03
+* ERM-1994 Prepare Plugin Find Agreement for RTL development
+* Updated stripes dependencies
+
+## 8.0.0 2021-10-06
+* Upgrade to stripes v7.
+* Bring search options in ui-plugin-find-agreement in line with ui-agreements search and filter. ERM-1600
+* Added descriptions to visible permission sets. ERM-1597
+* Included interface dependency for erm 5.0
+
+## 6.1.0 2021-06-16
+* ERM-1597 Add descriptions to visible permission set in ui-plugin-find-agreement
+
+## 6.0.0 2021-03-18
+* Upgrade to Stripes 6.0
+* Fixed issue with changing width. STCOM-749
+* Bumped okapiInterface dep to erm 4.0
+* Added devDep to stripes-cli
+
+## 5.0.0 2020-10-15
+* Upgrade to Stripes 5.0
+* Added tests for the plugin. ERM-1111
+* Fixed issue with date display. ERM-1153
+
+## 4.0.0 2020-06-10
+* Upgrade to Stripes 4.0
+* Added translations
+* Bumped the required node version to 10
+
+## 3.2.0 2020-03-11
+* Accessibility fixes
+* Added translations
+* Update to Stripes 3.0
+
+## 3.1.0 2019-12-02
+* Fixed clear button in search box.
+* Added support for agreement periods.
+* Added support for erm interface 2.0
+* Added permission set
+
+## 3.0.1 2019-04-06
+* Added translations
 
 ## 3.0.0 2019-07-22
 * Removed dependency on ui-agreements.